solar-tracker-dashboard/requirements.txt
Django==5.1.3
djangorestframework==3.14.0
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
